Moved graphql methods to new module
