+++ services/hype-backend/package.json
@@ -109,6 +109,7 @@
     "@taraxa-hype/config": "<rootDir>/config",
     "@taraxa-hype/user": "<rootDir>/src/modules/user",
     "@taraxa-hype/health": "<rootDir>/src/modules/health",
-    "@taraxa-hype/ipfs": "<rootDir>/src/modules/ipfs"
+    "@taraxa-hype/ipfs": "<rootDir>/src/modules/ipfs",
+    "@taraxa-hype/graphql": "<rootDir>/src/modules/graphql"
   }
 }
