+++ services/hype-backend/src/modules/graphql/index.ts
@@ -1,0 +1,2 @@
+export * from './graphQl.module';
+export * from './graphQl.service';
