+++ services/hype-backend/src/modules/reward/reward.module.ts
@@ -1,8 +1,6 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { GraphQLRequestModule } from '@golevelup/nestjs-graphql-request';
-import { ConfigModule, ConfigService } from '@nestjs/config';
 import { ScheduleModule } from '@nestjs/schedule';
 import { AuthModule } from '../auth';
 import { BlockchainModule } from '../blockchain';
@@ -12,20 +10,13 @@ import { RewardController } from './reward.controller';
 import { HypeReward } from '../../entities/reward.entity';
 import { RewardService } from './reward.service';
 import { PoolsController } from './pools.controller';
+import { GraphQlModule } from '../graphql';
 
 @Module({
   imports: [
     BlockchainModule,
     TypeOrmModule.forFeature([HypeReward, HypeClaim]),
-    GraphQLRequestModule.forRootAsync(GraphQLRequestModule, {
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: (config: ConfigService) => {
-        return {
-          endpoint: config.get<string>('auth.subGraphEndpoint'),
-        };
-      },
-    }),
+    GraphQlModule,
     ScheduleModule.forRoot(),
     AuthModule,
     HttpModule,
